requirements.txt
fastapi==0.111.0
numpy==1.26.4
pydantic==2.7.3
pyyaml==6.0.1
uvicorn==0.30.1